+++ core/client/fs/src/main/java/alluxio/client/file/FileInStream.java
@@ -14,6 +14,10 @@ package alluxio.client.file;
 import alluxio.Seekable;
 import alluxio.client.BoundedStream;
 import alluxio.client.PositionedReadable;
+import alluxio.exception.PreconditionMessage;
+import alluxio.util.io.BufferUtils;
+
+import com.google.common.base.Preconditions;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -26,6 +30,49 @@ import java.nio.ByteBuffer;
  */
 public abstract class FileInStream extends InputStream
     implements BoundedStream, PositionedReadable, Seekable {
+  private final byte[] mSingleByte = new byte[1];
+
+  @Override
+  public int read() throws IOException {
+    int bytesRead = read(mSingleByte);
+    if (bytesRead == -1) {
+      return -1;
+    }
+    Preconditions.checkState(bytesRead == 1);
+    return BufferUtils.byteToInt(mSingleByte[0]);
+  }
+
+  @Override
+  public int read(byte[] b) throws IOException {
+    return read(b, 0, b.length);
+  }
+
+  @Override
+  public int read(byte[] b, int off, int len) throws IOException {
+    Preconditions.checkNotNull(b, PreconditionMessage.ERR_READ_BUFFER_NULL);
+    return read(ByteBuffer.wrap(b), off, len);
+  }
+
+  /**
+   * Reads up to buf.remaining() bytes into buf. Callers should use
+   * buf.limit(..) to control the size of the desired read.
+   * <p>
+   * After a successful call, buf.position() will be advanced by the number
+   * of bytes read and buf.limit() should be unchanged.
+   * <p>
+   * In the case of an exception, the values of buf.position() and buf.limit()
+   * are undefined, and callers should be prepared to recover from this eventually.
+   * <p>
+   * Implementations should treat 0-length requests as legitimate, and must not
+   * signal an error upon their receipt.
+   *
+   * @param buf the ByteBuffer to receive the results of the read operation.
+   * @return the number of bytes read, possibly zero, or -1 if reach end-of-stream.
+   */
+  public int read(ByteBuffer buf) throws IOException {
+    return read(buf, buf.position(), buf.remaining());
+  }
+
   /**
    * Reads up to len bytes of data from the input stream into the byte buffer.
    *
@@ -52,25 +99,5 @@ public abstract class FileInStream extends InputStream
       byteBuffer.put(dest, 0, nread);
     }
     return nread;
-  }
-
-  /**
-   * Reads up to buf.remaining() bytes into buf. Callers should use
-   * buf.limit(..) to control the size of the desired read.
-   * <p>
-   * After a successful call, buf.position() will be advanced by the number
-   * of bytes read and buf.limit() should be unchanged.
-   * <p>
-   * In the case of an exception, the values of buf.position() and buf.limit()
-   * are undefined, and callers should be prepared to recover from this eventually.
-   * <p>
-   * Implementations should treat 0-length requests as legitimate, and must not
-   * signal an error upon their receipt.
-   *
-   * @param buf the ByteBuffer to receive the results of the read operation.
-   * @return the number of bytes read, possibly zero, or -1 if reach end-of-stream.
-   */
-  public int read(ByteBuffer buf) throws IOException {
-    return read(buf, buf.position(), buf.remaining());
   }
 }
