+++ core/client/fs/src/test/java/alluxio/client/file/MockFileInStream.java
@@ -62,11 +62,13 @@ public final class MockFileInStream extends FileInStream {
     mStream.skip(n);
   }
 
-  @Override public long getPos() throws IOException {
+  @Override
+  public long getPos() throws IOException {
     return mLength - remaining();
   }
 
-  @Override public int positionedRead(long position, byte[] buffer, int offset, int length)
+  @Override
+  public int positionedRead(long position, byte[] buffer, int offset, int length)
       throws IOException {
     throw new UnsupportedOperationException("positionedRead not implemented for mock FileInStream");
   }
